requirements.txt
django
gunicorn
whitenoise
django-environ
psycopg2-binary
djangorestframework
